Adc_Driver & Std_Types
Uploaded Adc Driver based on AUTOSAR specification, It's not completed yet but it's compiled no syntax erros.
uploaded Std_Types file based on AUTOSAR specs to be independent from the platform used.
only difference between this adc_driver and the old adc_driver is the Std_Types
and the configuration.